Print list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,14 +10,14 @@
  */
 size_t print_listint(const listint_t *h)
 {
-    unsigned int count = 0;
+	size_t count = 0;
 
-    while (h != NULL)
-    {
-        printf("%d\n", h->n);
-        h = h->next;
-        count++;
-    }
+	while (h != NULL)
+	{
+		printf("%d\n", h->n);
+		h = h->next;
+		count++;
+	}
 
-    return (count);
+	return (count);
 }
